bot/requirements.txt
discord.py>=2.3.2
aiohttp>=3.8.5
motor>=3.3.1
pymongo>=4.6.1
python-dotenv>=1.0.0
typing-extensions>=4.8.0